requirements.txt
langchain==0.3.25
langchain-google-vertexai==2.0.25
langchain-text-splitters==0.3.8
google-api-python-client==2.172.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.90.0
google-cloud-firestore==2.20.1
pypdf==5.6.0
python-dotenv==1.1.0
fastapi==0.115.12
uvicorn==0.34.2
